% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.iidm}
\alias{confint.iidm}
\title{Credible intervals for model parameters}
\usage{
\method{confint}{iidm}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted model object. Methods currently exist for the classes 
\code{"iidm"}, ... and ... .}

\item{parm}{a specification of which parameters are to be given credible 
intervals, either a vector of numbers or a vector of names. If missing, 
all parameters are considered.}

\item{level}{the credible level required.}

\item{...}{additional argument(s) for methods.}
}
\value{
A matrix (or vector) with columns giving the mean, lower and upper 
  credible limits for each parameter. These will be labelled as Estimate, 
  (1 - level)/2 and 1 - (1 - level)/2 in \% (by default 2.5\% and 97.5\%).
}
\description{
Computes credible intervals for one or more parameters in a 
  fitted model. Package \strong{spTReg} adds methods for \code{iidm} and ...
  fits.
}
\details{
\code{\link[stats]{confint}} is a generic function in package stats.
  
  These \code{confint} methods call the \code{\link[stats]{quantile}} 
  function, then find the credible intervals through the empirical quantiles
  of the posterior samples.
}
\author{
Jorge Castillo-Mateo
}
