% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetric_laplace.R
\name{Asymmetric-Laplace}
\alias{Asymmetric-Laplace}
\alias{dal}
\alias{qal}
\alias{pal}
\alias{ral}
\title{The Asymmetric Laplace Distribution}
\usage{
dal(x, mu = 0, sigma = 1, tau = 0.5, log = FALSE)

pal(q, mu = 0, sigma = 1, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

qal(p, mu = 0, sigma = 1, tau = 0.5, lower.tail = TRUE, log.p = FALSE)

ral(n, mu = 0, sigma = 1, tau = 0.5)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{log, log.p}{Logical. If \code{TRUE}, probabilities \eqn{p} are given as 
\eqn{\log(p)}.}

\item{lower.tail}{Logical. If \code{TRUE} (default), probabilities are
\eqn{P(X \le x)}, otherwise, \eqn{P(X > x)}.}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}

\item{size}{The Poisson binomial distribution has \code{size} times the
vector of probabilities \code{prob}.}

\item{prob}{Vector with the probabilities of success on each trial.}
}
\value{
\code{dal} gives the density, \code{pal} gives the distribution function,
  \code{qal} gives the quantile function and \code{ral} generates random 
  deviates.
  
  The length of the result is determined by ...
}
\description{
Density, distribution function, quantile function and random
  generation for the asymmetric Laplace distribution with location equal 
  to \code{mu}, scale equal to \code{sigma}, and asymmetry equal to 
  \code{tau}.
}
\details{
The asymmetric Laplace distribution with ... has density
  \deqn{f(x) = ...}
  for \eqn{x ...}.
  
  \eqn{p(x)} is computed using ... 
  
  The quantile is defined as ...
}
\references{
Kozumi H, Kobayashi G (2011). 
Gibbs sampling methods for Bayesian quantile regression. 
\emph{Journal of Statistical Computation and Simulation}, \strong{81}(11), 1565--1578. 
\doi{10.1080/00949655.2010.496117}

Yu K, Moyeed RA (2001). 
Bayesian quantile regression. 
\emph{Statistics & Probability Letters}, \strong{54}(4), 437--447.
\doi{10.1016/S0167-7152(01)00124-9}
}
\author{
Jorge Castillo-Mateo
}
