% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{deviance.iidm}
\alias{deviance.iidm}
\alias{deviance.spTm}
\title{Deviance}
\usage{
\method{deviance}{iidm}(object, value = c("mean", "samples"), ...)

\method{deviance}{spTm}(object, value = c("mean", "samples"), ...)
}
\arguments{
\item{object}{Object of class \code{"iidm"} or \code{"spTm"}.}

\item{value}{Model parameters used to compute the deviance (mean 
\eqn{D(\bar{\bm{\theta}})} or samples \eqn{D(\bm{\theta}^{(b)})}). Can be 
abbreviated.}

\item{...}{currently no additional arguments.}
}
\value{
The value of the deviance for each MCMC sample extracted from the 
  object \code{object}, or for the posterior mean of the MCMC samples 
  according to \code{value}.
}
\description{
Returns the deviance of a fitted model object \code{"iidm"} or
  \code{"spTm"}.
}
\details{
The deviance is defined according to
  \deqn{D(\bm{\theta}) = -2 \log([\mathbf{y} \mid \bm{\theta}]),}
  where \eqn{\mathbf{y}} are the data, \eqn{\bm{\theta}} are the unknown
  parameters of the model, and \eqn{[\mathbf{y} \mid \bm{\theta}]} is the 
  likelihood function.
  
  The deviance is only defined up to an additive constant. Different 
  constants have conventionally been used for different
  purposes and so \code{deviance} may give different values.
}
\references{
Spiegelhalter DJ, Best NG, Carlin BP, van der Linde A (2002). 
Bayesian measures of model complexity and fit (with discussion). 
\emph{Journal of the Royal Statistical Society, Series B}, \strong{64}(4), 583--639. 
\doi{10.1111/1467-9868.00353}
}
\seealso{
\code{\link{DIC}}, \code{\link{iidm}}, \code{\link{spTm}}
}
\author{
Jorge Castillo-Mateo
}
