% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIC.R
\name{DIC}
\alias{DIC}
\alias{DIC.default}
\title{Deviance information criterion}
\usage{
DIC(object, ...)

\method{DIC}{default}(object, ...)
}
\arguments{
\item{object}{Object of class \code{"iidm"} or \code{"spTm"}.}

\item{...}{currently no additional arguments.}
}
\value{
A numeric vector with three components corresponding to the DIC, the
  mean deviance \eqn{\overline{D(\bm{\theta})}}, and the effective number
  of parameters \eqn{p_{D}}.
}
\description{
Generic function calculating Deviance information criterion.
}
\details{
The DIC is calculated according to the formula 
  \deqn{\text{DIC} = \overline{D(\bm{\theta})} + p_{D},}
  where \eqn{D(\bm{\theta})} is the deviance, and 
  \eqn{p_{D} = \overline{D(\bm{\theta})} - D(\bar{\bm{\theta})}} is the 
  effective number of parameters. Bars represent the posterior mean.

  When comparing hierarchical models fitted by MCMC to the same 
  data, the smaller the DIC, the better the fit.
}
\references{
Spiegelhalter DJ, Best NG, Carlin BP, van der Linde A (2002). 
Bayesian measures of model complexity and fit (with discussion). 
\emph{Journal of the Royal Statistical Society, Series B}, \strong{64}(4), 583--639. 
\doi{10.1111/1467-9868.00353}
}
\seealso{
\code{\link{deviance}}, \code{\link{iidm}}, \code{\link{spTm}}
}
\author{
Jorge Castillo-Mateo
}
