% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krigeBayes.R
\name{krigeBayes}
\alias{krigeBayes}
\title{Bayesian kriging for spatially varying coefficients}
\usage{
krigeBayes(w, hp, coords, newcoords, ...)
}
\arguments{
\item{w}{a \eqn{B \times n} matrix with columns 
\eqn{w(\mathbf{s}_{1}),\ldots,w(\mathbf{s}_{n})}.}

\item{hp}{a \eqn{B \times 3} matrix with columns \eqn{\mu}, \eqn{\sigma},
and \eqn{\phi}.}

\item{coords}{a \eqn{n \times 2} matrix of the observation coordinates 
in \eqn{D \subset R^{2}}.}

\item{newcoords}{a \eqn{n_{0} \times 2} matrix of the new coordinates to 
krige, this is a grid \eqn{G_{D} \subset D \subset R^{2}}.}

\item{...}{additional argument(s) for methods.}
}
\value{
a \code{coda} object of posterior samples for the kriged 
  coefficient, that is \eqn{w(\mathbf{s}_{j})} for all
  \eqn{\mathbf{s}_{j} \in G_{D}}.
}
\description{
Bayesian kriging for a spatially varying coefficient in terms
  of the coefficients in the observation coordinates and the hyperparameters
  of the Gaussian process prior.

Function currently working but under development. 

\deqn{w(\mathbf{s}) \sim GP(\mu, C(\mathbf{s}, \mathbf{s}^{\prime}; \sigma^{2}, \phi))}
}
\author{
Jorge Castillo-Mateo
}
