% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.iidm}
\alias{summary.iidm}
\title{Summarizing model fits}
\usage{
\method{summary}{iidm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"iidm"}, usually the result of a call
to \code{\link{iidm}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The function \code{summary.iidm} computes and returns a list of 
  summary statistics of the fitted i.i.d. linear model given in 
  \code{object}, using the components (list elements) \code{"call"} and 
  \code{"terms"} from its argument, plus
  \item{r.squared}{if method="mean". A \code{coda} object of Bayesian \eqn{R^{2}} 
    (Gelman et al., 2018). The residual based \eqn{R^{2}} uses draws from 
    the residual distribution and is defined as 
    \deqn{R^{2,b} = \frac{Var_{\mu}^{b}}{Var_{\mu}^{b} + Var_{res}^{b}},}
    where
    \deqn{Var_{\mu}^{b} = V_{i=1}^{n} \hat{y}_{i}^{b}, \quad 
    Var_{res}^{b} = V_{i=1}^{n} \hat{e}_{i}^{b}.}
    The model based \eqn{R^{2}} uses draws from the modeled residual 
    variances and is defined as
    \deqn{Var_{res}^{b} = \sigma^{2,b}.}
  }
  \item{r.one}{if method="quantile". A \code{coda} object of Bayesian 
    \eqn{R^{1}(\tau)}. Instead of the residual sum of squares it uses the 
    residual weighted sum of absolute values associated with quantiles.}
}
\description{
\code{summary} method for class \code{"iidm"}.

Function currently not working.
}
\references{
Gelman A, Goodrich B, Gabry J, Vehtari A (2018). 
R-squared for Bayesian regression models.
\emph{The American Statistician}, \strong{73}(3), 307--309. 
\doi{10.1080/00031305.2018.1549100}.
}
\author{
Jorge Castillo-Mateo
}
