% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp.R
\name{tp}
\alias{tp}
\title{Time-varying coefficient in the formula}
\usage{
tp(x)
}
\arguments{
\item{x}{The explanatory variable for which time-varying coefficients are 
defined.}
}
\value{
An object of class \code{tpCoef}.
}
\description{
This function is used to define time-varying coefficients 
  within the formula for the space-time models.
}
\seealso{
\code{\link{spTm}}, \code{\link{sp}}
}
\author{
Jorge Castillo-Mateo
}
