% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation.spTm}
\alias{validation.spTm}
\title{Validation metrics for the space-time models}
\usage{
validation.spTm(model)
}
\arguments{
\item{model}{Object of class "spTm".}
}
\value{
\item{MSE}{Mean Squared Error.}
  \item{RMSE}{Root Mean Squared Error.}
  \item{MAE}{Mean Absolute Error.}
  \item{CRPS}{Continuous Rank Probability Score.}
  \item{CVG}{Coverage.}
  
  \item{WMAE}{Weighted Mean Absolute Error.}
  \item{p}{Probability of being under a quantile.}
}
\description{
This function is used to obtain validation metrics using MCMC
  samples.
}
\details{
Function currently not working.
}
\seealso{
\code{\link{spTm}}, \code{\link{predict.spTm}}
}
\author{
Jorge Castillo-Mateo
}
