% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iidm.R
\name{iidm}
\alias{iidm}
\title{MCMC sampling for the i.i.d. linear models}
\usage{
iidm(
  formula,
  data,
  subset,
  method = c("mean", "quantile"),
  quantile = 0.5,
  priors = list(beta = c(0, 1/100^2), sigma = c(2, 1)),
  starting = list(beta = 0, sigma = 1),
  n.samples = 1000,
  n.thin = 1,
  n.burnin = 1000,
  verbose = TRUE,
  n.report = 100,
  model = TRUE,
  x = FALSE,
  y = FALSE,
  ...
)
}
\arguments{
\item{formula}{an object of class \code{"\link[stats]{formula}"} (or one 
that can be coerced to that class): a symbolic description of the model to
be fitted. The details of model specification are given under ‘Details’.}

\item{data}{an optional data frame, list or environment (or object coercible 
by \code{\link{as.data.frame}} to a data frame) containing the variables 
in the model. If not found in data, the variables are taken from 
\code{environment(formula)}, typically the environment from which 
\code{iidm} is called.}

\item{subset}{an optional vector specifying a subset of observations to be 
used in the fitting process. (See additional details about how this 
argument interacts with data-dependent bases in the `Details' section of 
the \code{\link[stats]{model.frame}} documentation.)}

\item{method}{the method to be used for fitting, \code{"mean"} or 
\code{"quantile"} regression (default=\code{"mean"}).}

\item{quantile}{a numeric scalar containing the quantile level of interest 
(default=0.5).}

\item{priors}{a list with each tag corresponding to a parameter name. Valid 
tags are \code{"beta"} and \code{"sigma"}.}

\item{starting}{a list with each tag corresponding to a parameter name. Valid 
tags are \code{"beta"} and \code{"sigma"}.}

\item{n.samples}{the number of MCMC iterations after \code{n.burnin}.}

\item{n.thin}{the number of MCMC iterations kept is 1 out of \code{n.thin} 
iterations.}

\item{n.burnin}{the number of MCMC iterations discarded at the beginning.}

\item{verbose}{if \code{TRUE}, model specification and progress of the 
sampler is printed to the screen. Otherwise, nothing is printed to the 
screen.}

\item{n.report}{the interval to report MCMC progress.}

\item{model, x, y}{logicals. If \code{TRUE} the corresponding components of 
the fit (the model frame, the model matrix, the response) are returned.}

\item{...}{currently no additional arguments.}
}
\value{
An object of class \code{iidm}, which is a list comprising:
  \item{p.params.samples}{a \code{coda} object of posterior samples for the 
    model parameters, that is beta's and sigma.}
  \item{residuals}{a \code{coda} object of the residuals, that is response 
    minus fitted values.}
  \item{fitted.values}{the fitted mean or quantile values.}
  \item{method}{the method used.}
  \item{quantile}{if method="quantile", the quantile level used.}
  \item{xlevels}{(only where relevant) a record of the levels of the factors
    used in fitting.}
  \item{call}{the matched call.}
  \item{terms}{the \code{\link[stats]{terms}} object used.}
  \item{y}{if requested, the response used.}
  \item{x}{if requested, the model matrix used.}
  \item{model}{if requested (the default), the model frame used.}
}
\description{
This function is used to draw MCMC samples using the Gibbs 
  sampler that fits the Bayesian linear regression models with i.i.d. 
  errors.
}
\details{
The function \code{iidm} can fit two types of models, depending on whether
  the mean or the quantile is of interest.
  
  \item{Model:}{
  \deqn{Y_i = \mathbf{x}_i \bm{\beta} + \epsilon_i, \qquad i=1,\ldots,n,}}
  \item{Priors:}{}
}
\examples{
## Annette Dobson (1990) "An Introduction to Generalized Linear Models".
## Page 9: Plant Weight Data.
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)
set.seed(12345)
iidm.D9 <- iidm(weight ~ group, verbose = FALSE)
set.seed(12345)
iidm.D90 <- iidm(weight ~ group - 1, verbose = FALSE) # omitting intercept

}
\author{
Jorge Castillo-Mateo
}
