% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{sp}
\alias{sp}
\title{Space-varying coefficient in the formula}
\usage{
sp(x)
}
\arguments{
\item{x}{The explanatory variable for which space-varying coefficients are 
defined.}
}
\value{
An object of class \code{spCoef}.
}
\description{
This function is used to define space-varying coefficients 
  within the formula for the space-time models.
}
\seealso{
\code{\link{spTm}}, \code{\link{tp}}
}
\author{
Jorge Castillo-Mateo
}
