% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spTm.R
\name{spTm}
\alias{spTm}
\title{MCMC sampling for the space-time models}
\usage{
spTm(
  formula,
  u,
  v,
  x_alpha,
  data,
  subset,
  na.action,
  method = c("mean", "quantile"),
  quantile = 0.5,
  coords,
  priors = list(beta = c(0, 1/100), sigma = c(0.1, 0.1), rho = c(0, 1/100), hpGamma =
    list(sigma = c(0.1, 0.1), rho = c(0, 1/100)), hpAlpha = list(prec0 = 1/100, sigma =
    c(0.1, 0.1), phi = c(2, 100)), hpSigma = list(mu = 0, sigma = c(0.1, 0.1), phi = c(2,
    100)), hpRho = list(mu = 0, sigma = c(0.1, 0.1), phi = c(2, 100))),
  starting = list(beta = 0.01, sigma = 1, rho = 0, gamma = 0, alpha = 0, hpGamma =
    c(sigma = 1, phi = 3/100), hpAlpha = c(sigma = 1, phi = 3/100), hpSigma = c(mu = 0,
    sigma = 1, phi = 3/100), hpRho = c(mu = 0, sigma = 1, phi = 3/100)),
  n.samples = 1000,
  n.thin = 1,
  n.burnin = 1000,
  verbose = FALSE,
  n.report = 100,
  parallel = FALSE,
  n.threads = 0,
  model = TRUE,
  x = FALSE,
  y = FALSE,
  ...
)
}
\arguments{
\item{formula}{an object of class \code{"\link[stats]{formula}"} (or one 
that can be coerced to that class): a symbolic description of the model to
be fitted. The details of model specification are given under ‘Details’.}

\item{u, v}{(in addition to \code{formula}) \eqn{N \times q} and 
\eqn{N \times r} matrices of regression variables accompanying a 
time-varying coefficient and a spatially-varying coefficient, 
respectively.}

\item{data}{a data frame object with named columns giving the data to be 
fit; any explanatory variable necessary for modeling any of the 
parameters; and named columns \code{t}, \code{l}, and \code{s}, 
representing the numeric index of the data in the long and short 
temporal scales, and space, respectively; sorted by space. If not found 
in \code{data}, the variables are taken from \code{environment(formula)}, 
typically the environment from which \code{spTm} is called.}

\item{coords}{a \eqn{n \times 2} matrix of the observation coordinates in 
\code{R^2} (e.g., easting and northing).}

\item{priors}{a list with each tag corresponding to a parameter name.}

\item{starting}{a list with each tag corresponding to a parameter name.}

\item{n.samples}{the number of MCMC iterations after \code{n.burnin}.}

\item{n.thin}{the number of MCMC iterations kept is 1 out of \code{n.thin} 
iterations.}

\item{n.burnin}{the number of MCMC iterations discarded at the beginning.}

\item{verbose}{if \code{TRUE}, model specification and progress of the 
sampler is printed to the screen. Otherwise, nothing is printed to the 
screen.}

\item{n.report}{the interval to report Metropolis sampler acceptance and 
MCMC progress.}

\item{...}{currently no additional arguments.}

\item{location.fun, ar.fun, scale.fun}{(not working) an object of class 
\code{\link{formula}} describing a model for each 
parameter using columns from \code{data}. \code{data} must be supplied if 
any of these arguments have an explanatory variable. Three options are 
currently supported for \code{ar.fun}: the default \code{~ -1} with no 
autoregressive term, \code{~ 1} with a common autoregressive term, and 
\code{~ sp(1)} with a spatially-varying autoregressive term. Two options 
are currently supported for \code{scale.fun}: the default \code{~ 1} with 
a common scale term, and \code{~ sp(1)} with a spatially-varying scale 
term.}

\item{center.scale}{(not currently available) if \code{TRUE}, non-constant 
columns of \eqn{X} are 
centered on zero and scaled to have variance one. If 
\code{\link{predict.spTm}} is subsequently called this centering and 
scaling is applied automatically.}
}
\value{
An object of class \code{spTm}, which is a list comprising (this 
  list of outputs must be updated):
  \item{coords}{the \eqn{n x 2} matrix specified by \code{coords}.}
  \item{p.params.samples}{A \code{coda} object of posterior samples for the 
    defined parameters.}
  \item{acceptance}{The Metropolis sampling acceptance percent.}
  
  The return object will include additional objects used for subsequent 
  Gaussian process kriging, data kriging, and model fit evaluation using 
  \code{\link{predict.spTm}}, \code{\link{validation.spTm}}, 
  respectively.
}
\description{
This function is used to draw MCMC samples using the 
  Metropolis-within-Gibbs sampler that fits the Bayesian spatio-temporal
  models.

Function currently working but under development.
  
  Model:
  
  \deqn{\mathbf{Y}_{t\ell} = \mathbf{O}_{t\ell} + \mathbf{P} (\mathbf{Y}_{t,\ell-1} - \mathbf{O}_{t,\ell-1}) + \bm{\epsilon}_{t\ell},}
  \deqn{\mathbf{O}_{t\ell} = \mathbf{X}_{t\ell} \bm{\beta} + \mathbf{U}_{t\ell} \bm{\gamma}_{t} + \sum_{m=1}^{r} \mathbf{V}_{t\ell,m} \bm{\alpha}_{m},}
  \deqn{\bm{\epsilon}_{t\ell} \sim \text{i.i.d. } N_{n}(\mathbf{0}_{n}, \bm{\Sigma}).}

  Parallel execution is only available if the package was compiled with 
  \code{OpenMP} support. This mode samples the exponential latent variables 
  for quantile regression in parallel, but the current implementation does 
  not improve computational times. Parallel and sequential execution are
  both reproducible with \code{\link{set.seed}}, and the numerical results 
  between the two modes should be the same.
}
\references{
Castillo-Mateo J, Lafuente M, Asín J, Cebrián AC, Gelfand AE, Abaurrea J (2022).
Spatial modeling of day-within-year temperature time series: an examination of daily maximum temperatures in Aragón, Spain. 
\emph{Journal of Agricultural, Biological and Environmental Statistics}, \strong{27}(3), 487--505. 
\doi{10.1007/s13253-022-00493-3}.

Castillo-Mateo J, Asín J, Cebrián AC, Gelfand AE, Abaurrea J (2023).
Spatial quantile autoregression for season within year daily maximum temperature data. 
\emph{Annals of Applied Statistics}, \strong{17}(3), 2305--2325. 
\doi{10.1214/22-AOAS1719}
}
\seealso{
\code{\link{confint.spTm}}, \code{\link{predict.spTm}}
}
\author{
Jorge Castillo-Mateo
}
