% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.iidm}
\alias{summary.iidm}
\title{Summarizing model fits}
\usage{
\method{summary}{iidm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"iidm"}, usually, a result of a call 
to \code{\link{iidm}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{

}
\description{
\code{summary} method for class \code{"iidm"}.
}
\references{
Gelman A, Goodrich B, Gabry J, Vehtari A (2018). 
R-squared for Bayesian regression models.
\emph{The American Statistician}, \strong{73}(3), 307--309. 
\doi{10.1080/00031305.2018.1549100}.
}
\author{
Jorge Castillo-Mateo
}
